pyproject.toml
[project]
name = "scania-rpg-2-java"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.46",
    "langchain>=1.2.9",
    "langchain-aws>=1.2.2",
    "langchain-community>=0.4.1",
    "langchain-pinecone>=0.2.13",
    "pinecone>=7.3.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
